Remove 'multi-environment' registration
This feature was required as we migrated from Slough/Bristol to AWS,
with some services deployed to 'poke' or 'prod' expecting to discover
services registered in 'live' or 'cq3' (or vice-versa).

Now that the platform isn't split across Slough/Bristol and AWS, we
don't need a service to register in many environments at once.